requirements.txt
bs4
requests
pycountry
datetime
pandas
google-cloud-language
lxml
demjson
python-dotenv